Remove entirely cLog
