+++ src/gamestates/cSetupSkirmishState.cpp
@@ -14,7 +14,6 @@
 #include "gameobjects/players/brains/cPlayerBrainSkirmish.h"
 #include "gameobjects/players/brains/superweapon/cPlayerBrainFremenSuperWeapon.h"
 #include "gameobjects/players/cPlayer.h"
-// #include "utils/cLog.h"
 #include "utils/Log.h"
 #include "utils/Graphics.hpp"
 #include "utils/RNG.hpp"
