--- src/utils/cLog.cpp
@@ -1,1 +1,0 @@
-#include "utils/cLog.h"
